+++ tests/cli/test_handlers.py
@@ -1,5 +1,6 @@
 from pytest_mock import MockerFixture
 
+from kpops.api import setup_handlers
 from kpops.component_handlers import ComponentHandlers
 from kpops.component_handlers.kafka_connect.kafka_connect_handler import (
     KafkaConnectHandler,
@@ -7,8 +8,9 @@ from kpops.component_handlers.kafka_connect.kafka_connect_handler import (
 from kpops.component_handlers.schema_handler.schema_handler import SchemaHandler
 from kpops.component_handlers.topic.handler import TopicHandler
 from kpops.config import KpopsConfig, SchemaRegistryConfig
-from kpops.pipeline import Pipeline
 from tests.cli.resources.custom_module import CustomSchemaProvider
+
+HANDLER_MODULE = "kpops.api"
 
 MODULE = CustomSchemaProvider.__module__
 
@@ -18,12 +20,12 @@ def test_set_up_handlers_with_no_schema_handler(mocker: MockerFixture):
         kafka_brokers="broker:9092",
         components_module=MODULE,
     )
-    connector_handler_mock = mocker.patch("kpops.pipeline.KafkaConnectHandler")
+    connector_handler_mock = mocker.patch(f"{HANDLER_MODULE}.KafkaConnectHandler")
     connector_handler = KafkaConnectHandler.from_kpops_config(config)
     connector_handler_mock.from_kpops_config.return_value = connector_handler
 
-    topic_handler_mock = mocker.patch("kpops.pipeline.TopicHandler")
-    wrapper = mocker.patch("kpops.pipeline.ProxyWrapper")
+    topic_handler_mock = mocker.patch(f"{HANDLER_MODULE}.TopicHandler")
+    wrapper = mocker.patch(f"{HANDLER_MODULE}.ProxyWrapper")
     topic_handler = TopicHandler(wrapper)
     topic_handler_mock.return_value = topic_handler
 
@@ -33,7 +35,7 @@ def test_set_up_handlers_with_no_schema_handler(mocker: MockerFixture):
         topic_handler=topic_handler,
     )
 
-    actual_handlers = Pipeline.setup_handlers(config)
+    actual_handlers = setup_handlers(config)
 
     connector_handler_mock.from_kpops_config.assert_called_once_with(config)
 
@@ -51,16 +53,16 @@ def test_set_up_handlers_with_schema_handler(mocker: MockerFixture):
         schema_registry=SchemaRegistryConfig(enabled=True),
         kafka_brokers="broker:9092",
     )
-    schema_handler_mock = mocker.patch("kpops.pipeline.SchemaHandler")
+    schema_handler_mock = mocker.patch(f"{HANDLER_MODULE}.SchemaHandler")
     schema_handler = SchemaHandler.load_schema_handler(config)
     schema_handler_mock.load_schema_handler.return_value = schema_handler
 
-    connector_handler_mock = mocker.patch("kpops.pipeline.KafkaConnectHandler")
+    connector_handler_mock = mocker.patch(f"{HANDLER_MODULE}.KafkaConnectHandler")
     connector_handler = KafkaConnectHandler.from_kpops_config(config)
     connector_handler_mock.from_kpops_config.return_value = connector_handler
 
-    topic_handler_mock = mocker.patch("kpops.pipeline.TopicHandler")
-    wrapper = mocker.patch("kpops.pipeline.ProxyWrapper")
+    topic_handler_mock = mocker.patch(f"{HANDLER_MODULE}.TopicHandler")
+    wrapper = mocker.patch(f"{HANDLER_MODULE}.ProxyWrapper")
     topic_handler = TopicHandler(wrapper)
     topic_handler_mock.return_value = topic_handler
 
@@ -70,7 +72,7 @@ def test_set_up_handlers_with_schema_handler(mocker: MockerFixture):
         topic_handler=topic_handler,
     )
 
-    actual_handlers = Pipeline.setup_handlers(config)
+    actual_handlers = setup_handlers(config)
 
     schema_handler_mock.load_schema_handler.assert_called_once_with(config)
